extensions-project-starter/requirements.txt
pyyaml
dt-cli
requests
pyjson